# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestion des Itinéraires",
        "description": "Documentation pour l'API de gestion des itinéraires",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/itineraries": {
            "get": {
                "tags": [
                    "Itinéraires"
                ],
                "summary": "Récupérer la liste des itinéraires",
                "operationId": "86675d12475f84f785e7563fa8b8a0d9",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filtrer par catégorie",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "min_duration",
                        "in": "query",
                        "description": "Filtrer par durée minimale",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_duration",
                        "in": "query",
                        "description": "Filtrer par durée maximale",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des itinéraires récupérée avec succès"
                    },
                    "500": {
                        "description": "Erreur serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "Itinéraires"
                ],
                "summary": "Créer un nouvel itinéraire",
                "operationId": "9ddf03b1ccb2fb24c86fa810f0614b38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "duration",
                                    "destinations"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Voyage au Sahara"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Désert"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "destinations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "name": {
                                                    "type": "string",
                                                    "example": "Merzouga"
                                                },
                                                "lodging": {
                                                    "type": "string",
                                                    "example": "Camp du désert"
                                                },
                                                "places": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Itinéraire créé avec succès"
                    },
                    "422": {
                        "description": "Données invalides"
                    },
                    "409": {
                        "description": "Itinéraire déjà existant"
                    }
                }
            }
        },
        "/api/itineraries/{id}": {
            "get": {
                "tags": [
                    "Itinéraires"
                ],
                "summary": "Récupérer un itinéraire spécifique",
                "operationId": "491c9ddb477072a43a407e3d306bd81c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'itinéraire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itinéraire récupéré"
                    },
                    "404": {
                        "description": "Itinéraire non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Itinéraires"
                ],
                "summary": "Mettre à jour un itinéraire",
                "operationId": "6c2bd72ca2c9fccd61ef1297b4c0b1a4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'itinéraire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Voyage au Sahara Modifié"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Aventure"
                                    },
                                    "duration": {
                                        "type": "integer",
                                        "example": 7
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Itinéraire mis à jour"
                    },
                    "404": {
                        "description": "Itinéraire non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Itinéraires"
                ],
                "summary": "Supprimer un itinéraire",
                "operationId": "f65090001e6f5de3421bde54a3dd140f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de l'itinéraire",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Itinéraire supprimé"
                    },
                    "404": {
                        "description": "Itinéraire non trouvé"
                    }
                }
            }
        }
    },
...
}